feat(#385): portfolio positions (#386)
* refactor: #385: use `TableCell` component in positions instead of `Table`

* refactor: #385: move position table to entities

* refactor: #385: split the `helpers` file to the transaction entity

* refactor: #385: split withdraw and close function from position store

* refactor: #385: split position store into independent functions

* fix: #385: implement filters in getDisplayPositions

* feat: #385: apply positions table to the portfolio page

* feat: #385: add infinite scroll to positions table

* fix: lint
