+++ src/entities/position/index.ts
@@ -1,0 +1,3 @@
+export { PositionsTable } from './ui/table';
+export type { PositionsTableProps } from './ui/table';
+export { updatePositionsQuery } from './api/use-positions';
