+++ src/entities/position/ui/not-connected-notice.tsx
