+++ src/entities/position/ui/positions-current-value.tsx
@@ -1,16 +1,19 @@
-import React from 'react';
-import { useMarketPrice } from '../../model/useMarketPrice';
+import { useMarketPrice } from '@/pages/trade/model/useMarketPrice';
 import { ValueViewComponent } from '@penumbra-zone/ui/ValueView';
-import { LoadingCell } from './cell';
 import { pnum } from '@penumbra-zone/types/pnum';
-import { DisplayPosition } from '../../model/positions';
+import { Skeleton } from '@penumbra-zone/ui/Skeleton';
+import { DisplayPosition } from '../model/types';
 
 export const PositionsCurrentValue = ({ order }: { order: DisplayPosition['orders'][number] }) => {
   const { baseAsset, quoteAsset } = order;
   const marketPrice = useMarketPrice(baseAsset.asset.symbol, quoteAsset.asset.symbol);
 
   if (!marketPrice) {
-    return <LoadingCell />;
+    return (
+      <div className='w-12 h-4'>
+        <Skeleton />
+      </div>
+    );
   }
 
   if (order.direction === 'Buy') {
@@ -19,20 +22,22 @@ export const PositionsCurrentValue = ({ order }: { order: DisplayPosition['order
         valueView={pnum(quoteAsset.amount.toNumber(), quoteAsset.exponent).toValueView(
           quoteAsset.asset,
         )}
-        density='slim'
       />
     );
   }
 
   const computedValue = baseAsset.amount.toNumber() * marketPrice;
   if (!Number.isFinite(computedValue)) {
-    return <LoadingCell />;
+    return (
+      <div className='w-12 h-4'>
+        <Skeleton />
+      </div>
+    );
   }
 
   return (
     <ValueViewComponent
       valueView={pnum(computedValue, quoteAsset.exponent).toValueView(quoteAsset.asset)}
-      density='slim'
     />
   );
 };
