+++ src/entities/transaction/model/tx-to-id.ts
@@ -1,0 +1,12 @@
+import { Transaction } from '@penumbra-zone/protobuf/penumbra/core/transaction/v1/transaction_pb';
+import { PartialMessage } from '@bufbuild/protobuf';
+import { sha256Hash } from '@penumbra-zone/crypto-web/sha256';
+import { TransactionId } from '@penumbra-zone/protobuf/penumbra/core/txhash/v1/txhash_pb';
+
+/**
+ * Hashes the transaction and returns the transaction id using sha256
+ */
+export const txToId = (tx: Transaction | PartialMessage<Transaction>) =>
+  sha256Hash(tx instanceof Transaction ? tx.toBinary() : new Transaction(tx).toBinary()).then(
+    inner => new TransactionId({ inner }),
+  );
