+++ src/pages/portfolio/ui/closed-positions.tsx
@@ -1,3 +1,13 @@
+import { PositionState_PositionStateEnum } from '@penumbra-zone/protobuf/penumbra/core/component/dex/v1/dex_pb';
+import { PositionsTable } from '@/entities/position';
+
 export const PortfolioClosedPositions = () => {
-  return <div className='text-text-primary'>Closed positions</div>;
+  return (
+    <PositionsTable
+      stateFilter={[
+        PositionState_PositionStateEnum.CLOSED,
+        PositionState_PositionStateEnum.WITHDRAWN,
+      ]}
+    />
+  );
 };
