+++ src/pages/portfolio/ui/open-positions.tsx
@@ -1,3 +1,6 @@
+import { PositionsTable } from '@/entities/position';
+import { PositionState_PositionStateEnum } from '@penumbra-zone/protobuf/penumbra/core/component/dex/v1/dex_pb';
+
 export const PortfolioOpenPositions = () => {
-  return <div className='text-text-primary'>Open positions</div>;
+  return <PositionsTable stateFilter={[PositionState_PositionStateEnum.OPENED]} />;
 };
