+++ src/pages/trade/ui/history-tabs.tsx
@@ -1,10 +1,12 @@
 import { useState } from 'react';
 import { useAutoAnimate } from '@formkit/auto-animate/react';
+import { PositionState_PositionStateEnum } from '@penumbra-zone/protobuf/penumbra/core/component/dex/v1/dex_pb';
 import { Tabs } from '@penumbra-zone/ui/Tabs';
 import { Density } from '@penumbra-zone/ui/Density';
 import { Toggle } from '@penumbra-zone/ui/Toggle';
 import { Text } from '@penumbra-zone/ui/Text';
-import Positions from './positions';
+import { PositionsTable } from '@/entities/position';
+import { usePathToMetadata } from '../model/use-path';
 
 enum PositionsTabsType {
   MY_POSITIONS = 'MY_POSITIONS',
@@ -14,6 +16,7 @@ export const HistoryTabs = () => {
   const [parent] = useAutoAnimate();
   const [tab, setTab] = useState<PositionsTabsType>(PositionsTabsType.MY_POSITIONS);
   const [showInactive, setshowInactive] = useState(false);
+  const { baseAsset, quoteAsset } = usePathToMetadata();
 
   return (
     <div ref={parent} className='flex flex-col w-screen desktop:w-auto'>
@@ -33,7 +36,21 @@ export const HistoryTabs = () => {
         </label>
       </div>
 
-      <Positions showInactive={showInactive} />
+      <div className='p-4'>
+        <PositionsTable
+          base={baseAsset}
+          quote={quoteAsset}
+          stateFilter={
+            showInactive
+              ? [
+                  PositionState_PositionStateEnum.OPENED,
+                  PositionState_PositionStateEnum.CLOSED,
+                  PositionState_PositionStateEnum.WITHDRAWN,
+                ]
+              : [PositionState_PositionStateEnum.OPENED, PositionState_PositionStateEnum.CLOSED]
+          }
+        />
+      </div>
     </div>
   );
 };
