--- src/pages/trade/ui/positions/no-positions.tsx
@@ -1,11 +1,0 @@
-import { Text } from '@penumbra-zone/ui/Text';
-
-export const NoPositions = () => {
-  return (
-    <div className='p-5'>
-      <Text small color='text.secondary'>
-        No liquidity positions opened
-      </Text>
-    </div>
-  );
-};
